restructure tools/ folder
