+++ deeprankcore/neuralnets/foutnet.py
@@ -8,7 +8,7 @@ from torch_scatter import scatter_mean
 from torch_geometric.nn.inits import uniform
 from torch_geometric.nn import max_pool_x
 
-from deeprankcore.tools.community_pooling import get_preloaded_cluster, community_pooling
+from deeprankcore.models.community_pooling import get_preloaded_cluster, community_pooling
 
 
 class FoutLayer(torch.nn.Module):
