+++ deeprankcore/tools/classdiagrams.sh
