+++ deeprankcore/tools/plotgraph.py
@@ -2,7 +2,7 @@
 
 import os
 from h5xplorer.h5xplorer import h5xplorer
-from deeprankcore.h5x.h5x_menu import context_menu
+from deeprankcore.tools.visualization.h5x_menu import context_menu
 
 baseimport = os.path.dirname(
     os.path.abspath(__file__)) + "/baseimport.py"
