+++ deeprankcore/tools/target.py
@@ -3,10 +3,13 @@ import h5py
 import os
 import numpy as np
 from deeprankcore.domain import targettypes as targets
+from typing import Dict, Union
+from pdb2sql import StructureSimilarity
+
 
 
 def add_target(graph_path, target_name, target_list, sep=" "):
-    """Add a target to all the graphs contains in hdf5 files
+    """Add a target to all the graphs in hdf5 files
 
     Args:
         graph_path (str, list(str)): either a directory containing all the hdf5 files,
@@ -77,3 +80,46 @@ def add_target(graph_path, target_name, target_list, sep=" "):
 
         except BaseException:
             print(f"no graph for {hdf5}")
+
+def compute_targets(pdb_path: str, reference_pdb_path: str) -> Dict[str, Union[float, int]]:
+
+    """Compute targets (lrmsd, irmsd, fnat, dockq, bin_class, capri_class) and outputs them as a dictionary
+
+    Args:
+        pdb_path (path): path to the scored pdb structure
+        reference_pdb_path (path): path to the reference structure required to compute the different target
+
+    Returns: a dictionary containing values for lrmsd, irmsd, fnat, dockq, bin_class, capri_class
+    """
+
+    ref_name = os.path.splitext(os.path.basename(reference_pdb_path))[0]
+    sim = StructureSimilarity(pdb_path, reference_pdb_path)
+
+    scores = {}
+
+    # Input pre-computed zone files
+    if os.path.exists(ref_name + ".lzone"):
+        scores[targets.LRMSD] = sim.compute_lrmsd_fast(
+            method="svd", lzone=ref_name + ".lzone"
+        )
+        scores[targets.IRMSD] = sim.compute_irmsd_fast(
+            method="svd", izone=ref_name + ".izone"
+        )
+
+    # Compute zone files
+    else:
+        scores[targets.LRMSD] = sim.compute_lrmsd_fast(method="svd")
+        scores[targets.IRMSD] = sim.compute_irmsd_fast(method="svd")
+
+    scores[targets.FNAT] = sim.compute_fnat_fast()
+    scores[targets.DOCKQ] = sim.compute_DockQScore(
+        scores[targets.FNAT], scores[targets.LRMSD], scores[targets.IRMSD]
+    )
+    scores[targets.BINARY] = scores[targets.IRMSD] < 4.0
+
+    scores[targets.CAPRI] = 5
+    for thr, val in zip([6.0, 4.0, 2.0, 1.0], [4, 3, 2, 1]):
+        if scores[targets.IRMSD] < thr:
+            scores[targets.CAPRI] = val
+
+    return scores
