+++ tests/test_community_pooling.py
@@ -2,7 +2,7 @@ import unittest
 import torch
 import numpy as np
 from torch_geometric.data import Data, Batch
-from deeprankcore.tools.community_pooling import (
+from deeprankcore.models.community_pooling import (
     community_detection,
     community_pooling,
     community_detection_per_batch,
